# app/server/static/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Job Search Intelligence API",
    "description": "AI-powered career intelligence platform — job discovery, company analysis, skill management, resume generation, and career insights.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Development server" }
  ],
...
  "paths": {
...
    "/api/tech-stack": {
      "get": {
        "tags": ["Skills"],
        "summary": "Get visible skills with aliases and tags",
        "parameters": [
          { "name": "category", "in": "query", "schema": { "type": "string", "enum": ["technical", "engineering", "professional", "domain", "career"] } }
        ],
        "responses": { "200": { "description": "List of visible skills" } }
      },
      "post": {
        "tags": ["Skills"],
        "summary": "Add a new skill",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "level": { "type": "integer" },
                  "category": { "type": "string" },
                  "source": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Skill created" } }
      }
    },
    "/api/tech-stack/{id}": {
      "put": {
        "tags": ["Skills"],
        "summary": "Update a skill",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Skill updated" } }
      },
      "delete": {
        "tags": ["Skills"],
        "summary": "Permanently delete a skill and its aliases",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Skill deleted" } }
      }
    },
...
  }
}